Fix icon loading issue
The icon didn't load in Firefox on Android when adding the app to the
homescreen. In addition Chrome could not load the icons in the devtools
application manifest section.

I found that the manifest is served with the incorrect content type
(application/octet-stream). This is fixed, but didn't solve the problem.

I changed some icon sizes to align with other web apps. This didn't
solve the problem.

I finally found that removing the Cross-Origin-Resource-Policy header
helped. I don't know why that would be the case. It feels like a bug in
browsers or I don't understand how the header works. Removing for now,
until I figure this out.